move state machine to its own file fix no new line for RobotState class
